+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## 1.1.0 
 ### Features
-1. [#96](https://github.com/influxdata/influxdb-client-go/pull/96)  Authorization API
+1. [#100](https://github.com/influxdata/influxdb-client-go/pull/100)  HTTP request timeout made configurable
   
 ## 1.0.0 [2020-04-01]
 ### Core