+++ CHANGELOG.md
@@ -1,9 +1,9 @@
 ## 1.1.0 
 ### Features
 1. [#99](https://github.com/influxdata/influxdb-client-go/pull/99)  Organizations API and Users API
+1. [#96](https://github.com/influxdata/influxdb-client-go/pull/96)  Authorization API
 
 ## 1.0.0 [2020-04-01]
-
 ### Core
 
 - initial release of new client version