+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 1. [#99](https://github.com/influxdata/influxdb-client-go/pull/99)  Organizations API and Users API
 1. [#96](https://github.com/influxdata/influxdb-client-go/pull/96)  Authorization API
 
+### Doc
+1. [#101](https://github.com/influxdata/influxdb-client-go/pull/101) Added examples to API doc
+
 ## 1.0.0 [2020-04-01]
 ### Core