+++ CHANGELOG.md
@@ -1,5 +1,6 @@
-## 1.1.0 
+## 1.1.0
 ### Features
+1. [#100](https://github.com/influxdata/influxdb-client-go/pull/100)  HTTP request timeout made configurable
 1. [#99](https://github.com/influxdata/influxdb-client-go/pull/99)  Organizations API and Users API
 1. [#96](https://github.com/influxdata/influxdb-client-go/pull/96)  Authorization API